Add files via upload
Home page attempt. I was able to get my navbar to have a picture in the background but was not able to get the picture to extend all the way along with my navbar being a little weird and I am not too sure why. I have attempted the side scroll but it was not working for me even with a new file but I was able to get the logos in there.